+++ src/main/java/net/sf/jsqlparser/util/validation/validator/CreateIndexValidator.java
@@ -14,6 +14,7 @@ import net.sf.jsqlparser.statement.create.index.CreateIndex;
 import net.sf.jsqlparser.statement.create.table.Index;
 import net.sf.jsqlparser.util.validation.metadata.NamedObject;
 import net.sf.jsqlparser.util.validation.ValidationCapability;
+import net.sf.jsqlparser.util.validation.ValidationUtil;
 
 /**
  * @author gitmotte
@@ -30,7 +31,7 @@ public class CreateIndexValidator extends AbstractValidator<CreateIndex> {
          // TODO validate for not existing ?? this may be a little bit more complex
             // because database-names share one space in most databases
             //            validateNameNotExists(c, NamedObject.index, index.getName());
-            validateOptionalColumnNames(ValidatorUtil.concat(tableFqn, index.getColumnsNames()), c);
+            validateOptionalColumnNames(ValidationUtil.concat(tableFqn, index.getColumnsNames()), c);
         }
     }
