--- src/test/java/net/sf/jsqlparser/util/validation/validator/ValidatorUtilTest.java
@@ -1,13 +1,0 @@
-package net.sf.jsqlparser.util.validation.validator;
-
-import org.junit.Test;
-import net.sf.jsqlparser.util.validation.ValidationTestAsserts;
-
-public class ValidatorUtilTest extends ValidationTestAsserts {
-
-    @Test
-    public void test() {
-
-    }
-
-}
